CRUD FINAL

+++ src/main/java/com/cursojava/curso/service/UsuarioServiceAPI.java
@@ -3,7 +3,7 @@ package com.cursojava.curso.service;
 import com.cursojava.curso.commons.GenericServiceAPI;
 import com.cursojava.curso.model.Usuario;
 
-public interface UsuarioServiceAPI extends GenericServiceAPI<Usuario, Long> {
+public interface UsuarioServiceAPI extends GenericServiceAPI<Usuario, Integer> {
 
     Usuario login(String correo, String clave);
